requirements.txt
aiohttp==3.11.18
aiosignal==1.3.2
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
Flask==3.1.0
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
packaging==24.2
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
